Add a toolchain for building for ios7.
Doesn't actually build with xcode build (I found you can make xcode build
configurations with cmake!) But the cmake command succeeds which is progress.

Leaving out the changes to libgit2-iOS.podspec for now because it not building
feels like going backwards.

It seems to fail because I didn't use the correct architecture and the
dependencies aren't in the project which means I probably need to make some
more cocoapods for libz and openssl.