Completely misunderstood how cocoapods work.
I mean I knew it copied the source over and what not but that never crossed my
mind when I was dealing with this whole prepare_command stuff. Apparently need to
get this building in the projects that include it not before hand.

Move ultimate header file location in a start to doing this.