Only compile .c files and removed deleted dirs.
Removing deleted dirs so we can use libgit2 as a submodule.
The errors I'm getting now seem to confirm my suspicion that the library wasn't
really being built previously. The project's still not building because some of the
headers are still being over ridden because the folder mapping isn't being
preserved.